AI Creation
Added a simple ai that will detect objects that enter its trigger sphere
and gives chase with a min and max range. Also has the ability to fire
weapons. However the Ai script seems to have trouble calling other
scripts and so the team check doesn't work and the firing does no damage
at the moment.